requirements.txt
attrs==20.3.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
importlib-metadata==2.1.1
iniconfig==1.1.1
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.4
nose==1.3.7
numpy==1.18.5
opencv-python==4.4.0.42
packaging==20.9
pandas==0.24.2
pathlib2==2.3.5
Pillow==7.2.0
pip==20.3.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
setuptools==20.7.0
six==1.15.0
spherecluster==0.1.7
toml==0.10.2
torch==1.4.0+cu92
torchvision==0.5.0+cu92
UNKNOWN==0.0.0
wheel==0.29.0
zipp==1.2.0
